Refactor private methods

+++ numberbaseball/src/main/java/com/github/lette1394/numberbaseball/Result.java
@@ -6,6 +6,6 @@ import lombok.Value;
 @Value
 @Builder
 public class Result {
-  int strike;
-  int ball;
+  long strike;
+  long ball;
 }
